Refactor coordinates compoents to Typescript
